pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "generic-llm-lib"
version = "0.3.1"
description = "A flexible, async-first agent library for LLM providers (Google Gemini, etc.) designed for scalable chatbots and server-side applications."
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "google-genai",
    "jsonref>=1.1.0",
    "mcp>=1.26.0",
    "openai",
    "pydantic>=2.0",
]

[project.urls]
"Homepage" = "https://github.com/NickBelaneDev/generic-llm-lib"

[dependency-groups]
dev = [
    "black>=24.0",
    "mypy>=1.0",
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.1.0",
]

[tool.pytest.ini_options]
pythonpath = "src"
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.black]
line-length = 120
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
mypy_path = "src"
explicit_package_bases = true
namespace_packages = true
